Update Rakefile and framework

+++ Rakefile
@@ -1,57 +1,59 @@
-# Lazy hard-coding of problem slugs
-SLUGS = %w{freq streets war password}
-# Number of files to be generated for the server
-TEST_FILES = 50
-# Number of test cases to include per file (including edge cases)
-# We set this to 200 for the competition
-TEST_CASES = 20
+require 'yaml'
+
+config = YAML.load_file('config.yml')
+env = 'devel'
+
+slug_dir = lambda do |slug, rest|
+  File.join('.', config['semester'], slug, rest)
+end
 
 task :default => [:test]
 
 task :build => [:gen, :test]
 
 task :gen, :slug do |t, args|
-  puts "Generating #{ TEST_FILES } files each with #{ TEST_CASES } cases"
-  SLUGS.each do |slug|
+  puts "Generating #{ config['test_files'] } files each with #{ config['test_cases'][env] } cases"
+  config['problems'].each do |slug, _|
     if !args[:slug] || args[:slug] == slug
+      require slug_dir.call(slug, 'problem.rb')
+      cls = ProgComp.const_get(slug.capitalize)
+      problem = cls.new
       # Read the edge file
       # We had a problem here where the edge file was not being updated to
       # include the correct number of inputs (aka human error), so people who
       # ignored the number of inputs line got too many solutions.  We need to
       # foolproof this more.
-      edge = File.open("#{ slug }/edge.in").read.gsub(/^# .*/, '').gsub(/^\s*$\n/, '')
-      count = edge.lines.first.to_i
-      edge = edge.lines.drop(1).join
+      edge = File.open(slug_dir.call(slug, 'edge.in')).read.gsub(/^# .*/, '').gsub(/^\s*$\n/, '') rescue nil
+      if edge
+        count = edge.lines.first.to_i
+        edge = edge.lines.drop(1).join
+      end
 
       # Make the destination folder
       prefix = "grader/#{ slug }/"
       mkpath prefix
-      require "./#{ slug }/gen.rb"
 
       print "Generating files for #{ slug }"
-      TEST_FILES.times do |i|
+      config['test_files'].times do |i|
         print "."
         file = "#{ prefix }/#{ i }"
         # Generate the input file
         File.open("#{ file }.in", 'w+') do |input|
           # Write the count
-          input.puts TEST_CASES
+          input.puts config['test_cases'][env]
 
           # Write the edge cases
-          input.write edge
+          input.write edge if edge
 
           # Add in our own from gen.rb
-          Generator.generate(TEST_CASES - count, :without_leader => true) do |line|
+          problem.generate(config['test_cases'][env] - (count or 0), :without_leader => true) do |line|
             input.puts line
           end
         end
 
         # Generate the corresponding output file
         File.open("#{ file }.out", 'w+') do |output|
-          # Get our solution
-          require "./#{ slug }/solution.rb"
-
-          Solver.solve(File.open("#{ file }.in")) do |line|
+          problem.solve(File.open("#{ file }.in")) do |line|
             output.puts line
           end
         end
@@ -65,28 +67,26 @@ require 'open3'
 module Open3
   # A more encapsulated function that takes a command, optional file to be
   # written to stdin, and a block that stdout is read and passed to.
-  def self.piped_input cmd, file=nil, &block
+  def self.piped_input cmd, file=nil
     popen3(cmd) do |stdin, stdout, stderr|
       stdin.write File.open(file).read if file rescue nil
-      # stderr raises an exception and prints an error. This could be reworked,
-      # since if you rescue the exception you wouldn't want the extra printing.
       unless stderr.eof?
-        puts "Error in `#{ cmd } < #{ file }`:"
-        print stderr.read
-        raise :stderr
+        raise Tester::TestError, "Error in `#{ cmd } < #{ file }`:\n#{ stderr.read }"
       end
-      yield stdout.read if block
+      yield stdout.read if block_given?
     end
   end
 end
 
 module Tester
-  TESTS = {
-    'freq'     => 'freq_Varun.py',
-    'war'      => '1DWar_Varun.py',
-    'password' => 'Main.java',
-    'streets'  => 'Main.java',
-  }
+  class DriverError < RuntimeError; end
+  class TestError < RuntimeError; end
+
+  def self.make(runner)
+    driver = constants.map{|c| const_get c }.find{|cls| cls.respond_to? :test_on? and cls.test_on? runner }
+    raise DriverError, "No driver available" unless driver
+    driver.new runner
+  end
 
   class Base
     def test input
@@ -97,11 +97,10 @@ module Tester
         end
 
         # Diff with given output
-        Open3.piped_input("colordiff -u #{ input.gsub(/in$/, 'out') } .test.diff") do |diffout|
+        Open3.piped_input("diff -u #{ input.gsub(/in$/, 'out') } .test.diff") do |diffout|
           unless diffout.empty?
-            # Diff error! Same thing here with the printing and raising.
-            print diffout
-            raise "Test failed on #{ input }"
+            # Diff error!
+            raise TestError, "Test failed on #{ input }\n#{ diffout }"
           end
         end
       end
@@ -117,6 +116,10 @@ module Tester
       # Compile the java file before the tests run
       Open3.piped_input("javac #{ file }")
     end
+
+    def self.test_on? filename
+      filename =~ /\.java$/
+    end
   end
 
   class Python < Base
@@ -124,39 +127,48 @@ module Tester
       @file = file
       @cmd = 'python'
     end
+
+    def self.test_on? filename
+      filename =~ /\.py$/
+    end
   end
 end
 
 task :test, :slug do |t, args|
-  # Try out Varun's solutions
-  SLUGS.each do |slug|
+  # Try out other solutions
+  config['problems'].each do |slug, cfg|
     if !args[:slug] || args[:slug] == slug
       # Testing :slug
-      print "Testing #{ slug }"
-      # Make sure there is a test defined
-      test = Tester::TESTS[slug]
-      next unless test
+      puts "Testing #{ slug }"
 
-      # Check file type
-      ext = File.extname(test)
-      case ext
-      when '.java'
-        tester = Tester::Java.new("#{ slug }/#{ test }")
-      when '.py'
-        tester = Tester::Python.new("#{ slug }/#{ test }")
-      end
+      # Iterate over each runner
+      runners = cfg['runners']
+      next unless runners
+      runners.each do |runner|
+        begin
+          print "  #{ runner } "
+          tester = Tester::make slug_dir.call(slug, runner)
 
-      # Do for each file
-      TEST_FILES.times do |i|
-        tester.test "grader/#{ slug }/#{ i }.in"
-        print "."
+          # Do for each file
+          config['test_files'].times do |i|
+            tester.test "grader/#{ slug }/#{ i }.in"
+            print "."
+          end
+          puts "done"
+        rescue Tester::DriverError => e
+          puts e
+          next
+        rescue Tester::TestError => e
+          # Exit cleanly
+          puts e
+          exit
+        end
       end
-      puts "done"
     end
   end
 
   # If we get this far, everything succeeded and we can delete this temp file.
   # We don't really want to delete otherwise, since it can be useful for
   # debugging.
-  rm '.test.diff'
+  rm '.test.diff' if File.exists? '.test.diff'
 end
